Optimise pretty group rendering, prepare tests
